feat(ssr): provide `createAlgoliaSSRClient()`
